NitroSniper/requirements.txt
colorama==0.4.6
pystyle==2.9
tls-client==1.0.1
typing_extensions==4.12.2
websocket-client==1.8.0
